PrOH-Modeller/Pro-python/requirements.txt
pip
Flask
nltk
python-pptx
spacy
pandas
openpyxl
gunicorn

